COMP: Fix warning about conversion from double to unsigned long.

+++ Testing/Code/Review/itkBinaryShapeKeepNObjectsImageFilterTest1.cxx
@@ -58,7 +58,7 @@ int itkBinaryShapeKeepNObjectsImageFilterTest1(int argc, char * argv[])
   TEST_SET_GET_VALUE( BackgroundValue, KeepNObjects->GetBackgroundValue() );
 
   //testing get and set macros for Lambda 
-  double numberOfObjects = atof( argv[5] );
+  double numberOfObjects = atoi( argv[5] );
   KeepNObjects->SetNumberOfObjects( numberOfObjects );
   TEST_SET_GET_VALUE( numberOfObjects, KeepNObjects->GetNumberOfObjects() );
